When a pattern variable is defined on only one side of a disjunction... (#100)
* Adjust tests to reflect changes in Julia 1.11 nightlies.
* When a pattern variable is defined on only one side of a disjunction,
it is now an error to use it again later in the pattern,
or in a value expression. The reason is that the semantics
are too confusing. If you get this error, just rename the
pattern variable to avoid the conflict, or use a wildcard
instead.

This is technically a breaking change, but I'm guessing
that it will occur only rarely in practice. Therefore I
am only bumping the minor version number.

Fixes #99